Report whether a group existed prior to deletion
Previously, a Delete request on a group always succeeded, regardless of
whether the group actually existed.

Now, the Store API has been updated to allow for indication of whether
the group *actually* existed in the store before trying to delete it.
This allows us to report to the user whether they, for example, mistyped
the group name.

Note that while the DynamoDB store has been updated to satisfy the new
interface, it does not actually report the new information, and
maintains the previous behavior of acting as if the group always
existed. I need to look more into the DynamoDB API to figure out how to
get the information I want.

+++ pkg/store/bbolt/bbolt.go
@@ -88,13 +88,19 @@ func (b Store) Put(name string, options []string) error {
 }
 
 // Delete removes the named group from the store.
-func (b Store) Delete(name string) error {
-	return b.db.Update(func(tx *bolt.Tx) error {
+func (b Store) Delete(name string) (existed bool, err error) {
+	err = b.db.Update(func(tx *bolt.Tx) error {
 		bucket := tx.Bucket([]byte(b.bucket))
 		if bucket == nil {
-			return errors.Errorf("group %q does not exist", name)
+			return nil
 		}
 
+		if bucket.Get([]byte(name)) == nil {
+			return nil
+		}
+
+		existed = true
 		return errors.Wrapf(bucket.Delete([]byte(name)), "deleting group %q", name)
 	})
+	return
 }
